removed persistent cookie functionality until it works properly
Couldn't resolve error in streamlit-cookies-manager library due to outdated use of st.cache bc of old streamlit version, so decided to remove this functionality for now rather than have users see this warning when visiting the site. Need to find permanent solution to allow users to stay logged in across browser sessions. This was a challenge with Google Auth as well.